refactor: user service interface 추출
