+++ app/src/main/java/com/techcourse/service/AppUserService.java
@@ -1,0 +1,46 @@
+package com.techcourse.service;
+
+import javax.sql.DataSource;
+import com.interface21.transaction.support.JdbcTransactionManager;
+import com.interface21.transaction.support.JdbcTransactionTemplate;
+import com.techcourse.dao.UserDao;
+import com.techcourse.dao.UserHistoryDao;
+import com.techcourse.domain.User;
+import com.techcourse.domain.UserHistory;
+
+public class AppUserService implements UserService {
+
+    private final UserDao userDao;
+    private final UserHistoryDao userHistoryDao;
+    private final JdbcTransactionTemplate jdbcTransactionTemplate;
+
+    public AppUserService(
+            final UserDao userDao,
+            final UserHistoryDao userHistoryDao,
+            final DataSource dataSource
+    ) {
+        this.userDao = userDao;
+        this.userHistoryDao = userHistoryDao;
+        this.jdbcTransactionTemplate = new JdbcTransactionTemplate(new JdbcTransactionManager(dataSource));
+    }
+
+    @Override
+    public User findById(final long id) {
+        return userDao.findById(id);
+    }
+
+    @Override
+    public void insert(final User user) {
+        userDao.insert(user);
+    }
+
+    @Override
+    public void changePassword(final long id, final String newPassword, final String createBy) {
+        jdbcTransactionTemplate.execute((transaction -> {
+            final var user = userDao.findById(id);
+            user.changePassword(newPassword);
+            userDao.update(user);
+            userHistoryDao.log(new UserHistory(user, createBy));
+        }));
+    }
+}
