+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,43 +1,12 @@
 package com.techcourse.service;
 
-import javax.sql.DataSource;
-import com.interface21.transaction.support.JdbcTransactionManager;
-import com.interface21.transaction.support.JdbcTransactionTemplate;
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
 
-public class UserService {
+interface UserService {
 
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
-    private final JdbcTransactionTemplate jdbcTransactionTemplate;
+    User findById(long id);
 
-    public UserService(
-            final UserDao userDao,
-            final UserHistoryDao userHistoryDao,
-            final DataSource dataSource
-    ) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-        this.jdbcTransactionTemplate = new JdbcTransactionTemplate(new JdbcTransactionManager(dataSource));
-    }
+    void insert(User user);
 
-    public User findById(final long id) {
-        return userDao.findById(id);
-    }
-
-    public void insert(final User user) {
-        userDao.insert(user);
-    }
-
-    public void changePassword(final long id, final String newPassword, final String createBy) {
-        jdbcTransactionTemplate.execute((transaction -> {
-            final var user = userDao.findById(id);
-            user.changePassword(newPassword);
-            userDao.update(user);
-            userHistoryDao.log(new UserHistory(user, createBy));
-        }));
-    }
+    void changePassword(long id, String newPassword, String createBy);
 }
